New option company-abort-manual-when-too-short
For manually started completions, backspacing to a prefix shorter
than company-minimum-prefix-length still shows completions.

Setting the new option to t turns that off, so that completions
disappear whenever your prefix gets too short.

With either value of the new option, even if the user changes the
selection, that no longer suffices to make any prefix valid in this
completion invocation.