clang doc-buffer implementation #288
fetch a candidate's documentation directly from its source file

Use the current context to resolve the right candidate in the AST dump, then
locate the documentation in the candidate's source file.

The purpose of the current context is to disambiguate similar results in the AST
dump, such as when the candidate is a member of a data structure and in the dump
there are other possible candidates with the same format (members of other data
structures or variables with the same declaration).