Fix #474 by filtering dabbrev completions at the end
It seems impossible to combine supporting company-dabbrev-char-regexp, with
arbitrary values, with using the "word boundary anchor" (or any of the three
relevant anchors). And without anchors, using the regexp engine without
lookaheads or etc, we must filter completions at the end.

Which is okay, because it can allow us to cache the full list globally and
refresh it less often, as well as implement fuzzy filtering more easily (#418).