Ignore eclim when an eclipse proj does not exist.
The ~company-eclim~ completion backend is triggered inside Java files,
but the function ~company-eclim--project-dir~ fails and throws a
backtrace if the Java file is not part of an Eclipse project.

This happens because the call to ~locate-dominating-file~ returns nil
when such a file does not exist, and ~expand-file-name~ expects a string
input.

This commit fixes the error path by explicitly checking that
~locate-dominating-file~ returns a non-nil value.