Fixed company-semantic-annotation, so that it ..
.. can appropriately deal with function pointers in C.

As an example of the original problem, consider the struct

   struct blah {
       int (*hates)(uint8_t, uint32_t);
       void *blank;
   }

the annotation returned by company-semantic-annotation would be
         (*hates)(uint8_t, uint32_t)
which when coupled with the default value of 'company-semantic-insert-arguments'
yields decidedly annoying results. 

This patch fixes the problem. Additionally. you can find a gallery 
of screenshots illustrating the problem http://imgur.com/gallery/pyxwg

The issue with the original code seems to be the implicit assumption that the 
type of blah::hates is not seen as a function by semantic. This is incorrect, as 
semantic properly designates it a function-pointer which being callable, makes it 
a function.