Take the exact candidate as an nonoptional arg
By this change `company-preview-show-at-point`
takes the exact completion candidate
as an nonoptional arg. So the candidate was added as
the arg where this function is called.

At the previous change `company-preview-frontend` was modified to
take the new argument, but it was returned at this time.
Made `company-preview-common-frontend` not to call
`company-preview-frontend`, and instead it called
`company-preview-show-at-point` directly.