company-tng: default config: Disable company-require-match
Requiring a match doesn't work with company-tng because the expected behavior is
to be able to type any key during completion and it will be inserted into the
buffer.

The default value for company-require-match is to require match on explicit
action, so we can reproduce as following:

1. Open *scratch*
2. Type "(de"
3. M-x company-complete
4. RET RET RET...

The RET won't register because it's not part of the completion but this is not
how company-tng should work. The RETs should be inserted into the buffer.