Prevent unnecessary and unexpected calls to capf hooks
In company--capf-post-completion, we have to access cache directly because,
according to company--capf-data, it is already stale.  Indeed, the freshly
inserted completion has changed point, but the cache is not really stale for
getting the needed :exit-function.  Therefore, we use this cache to prevent a
second, unexpected call to the capf hooks.

* company-capf.el (company--capf-post-completion): Access cache
directly.