company: company--company-command-p: Fix lookup-key
Make sure it returns an actual command.

`lookup-key' returns a number when the key sequence is longer than 1 but it
isn't a valid sequence of prefix characters. The number evaluates to true but
the key isn't in the keymap.

When that happens the completion gets lost. To reproduce, complete a variable
name and type C-x C-s to save.

Fixes #857

Reported-by: Thomas Fini Hansen <https://github.com/xendk>