Set tooltip background color for terminals with less than 88 colors
The current implementation specifies the foreground color for tooltips to be black
but does not gives the background color on terminals with less color capability,
which causes invisible text when the background is black on <88-color terminals.

This patch fixes the problem by giving a default background color.
This change does not affect >=88-color terminals.