EUDC Back-End
This new back-end collects completion candidates for email addresses from EUDC,
and offers them when editing email message header fields that take one or more
email addresses (To, Cc, Bcc, From, and Reply-To) in message modes (configurable
via company-eudc-modes; default are message-mode, notmuch-message-mode, and
mail-mode).

Since EUDC queries may involve multiple remote servers, they can take many
seconds to complete. Hence, two options for using this back-end are provided,
and you must decide which option to use.

The first option adds the back-end to the global list of back-ends, so that EUDC
is queried whenever company-mode attempts completion in a suitable email message
header field. This option is intended for when the additional delay induced by
the EUDC query seems acceptable.

The second option is intended for when the additional delay induced by the EUDC
query does not seem acceptable, and the user therefore will want EUDC to be
queried only upon explicit request. To achieve this, the second option offers a
function that can be bound to a key chord (e.g. C-tab).