Can not search for more than one word when completing via EUDC
Previously, only the word preceding point (company-grab-symbol) would be taken
and searched for with EUDC. This is now replaced with a more sophisticated
look-back, which captures several words, up to the next colon (':') or
comma (',') preceding point.

This means, that users can now search for full names (e.g. first and last),
which is more in line with what EUDC users are used to.

Thanks to @alexmurray for suggesting this improvement in
https://github.com/company-mode/company-mode/pull/997#discussion_r467689727