Fix company-gtags to work with tramp and multiple locations.
* company-gtags.el (company-gtags-executable) : Now default to nil
because the search is made locally if not set.
(company-gtags--executable-p) : Function to search and cache the global
executable locally in the buffer.
(company-gtags--annotation) : Use the new function
company-gtags--executable-p instead of the custom variable
(company-gtags--fetch-tags) : Use the new function
company-gtags--executable-p but execute the process-file from the caller
buffer to have all the context. The output is written back to the
temporal buffer.