company-text-icons-margin: Always adjust bg on select
Before when users specified a BG entry in `company-text-icons-mapping`
it either had to be a face, a string or a CONS of the above to indicate
variants for when the item was selected and when it wasn't. If the CONS
wasn't given then the same background color was used both for selected
and unselected items. With this patch the BG is now always adjusted in
this case. If for example you add `(array "a" fg-face "red")` to
`company-text-icons-mapping` then when the tooltip selects an array
entry the foreground will now change slightly to indicate it's selected.
If you want to maintain the previous behaviour you can set the BG entry
to a cons of the same color (eg: `(array "a" fg-face ("red" . "red"))`).