company--multi-backend-adapter: Consider min prefix in 'candidates' separately
This should help with
https://github.com/company-mode/company-mode/discussions/1398,
https://github.com/company-mode/company-mode/issues/641,
https://github.com/company-mode/company-mode/issues/553,
https://github.com/company-mode/company-mode/issues/485,
https://github.com/company-mode/company-mode/pull/1147,
https://github.com/company-mode/company-mode/issues/840

With some of the above already closed, and some describing more serious
scenarios than others.

Anyway, this is a halfway step to dropping framework-level caching (or moving it to
backends).  For now this seems like an adequate solution, given that the most
"heavy" backend these days is `company-capf`, which disables caching already.
Though `company-dabbrev-code` is a close next.