company-dabbrev: Fix mismatched type in company-dabbrev--fetch
This fixes a regression introduced in 374322cea.

Before that commit, company-dabbrev--fetch calls company-dabbrev--search which
returns a hash table. Then company-dabbrev--filter converts it to a list (hash
table's keys), then (mapcar 'downcase words) is applied to the list.

After that commit, if "downcase-p" is set then the hash table is passed directly
to (mapcar 'downcase words) which produces an error "Wrong type argument:
sequencep, #s(hash-table ..."

This commit changes the return type of company-dabbrev--fetch from a hash table
to a list, although that doesn't seem to break other functionality in quick
testing.