company--common-or-matches: Synchronize with 'match' more
This is basically a reversal of the fix for #519, which now makes sense since
`expand-common` got closer to `match` (fuzzy, non-prefix expansions).

Highlighting is not predictive anymore - we don't highlight the common part
which would be inserted, at least with backends that support `match`.

Basically, CAPF's `try-completion` does not give us enough info (even if we
didn't mind scanning the collection once more during every update) to overlay
the "expanded common" string over the completions in the popup.