company-capf--candidates: Abort when the underlying capf changes
That cancels completion, forcing the re-fetching of data (after the idle delay).

Making sure that completions from one capf aren's displayed where another capf
is in force (when that choice is done based on buffer contents). This is
probably only relevant with "asynchronous" sources, where the frontends are
repainted in the middle of fetching the updated completions.