company-finish: Restart completion if ended up in new field
This seems necessary for good file completion experience (together with
"boundaries" support).  Both with company-files and completion in 'M-x shell'.

We run post-completion and then try begin-new again after.  But only if
post-completion didn't do any changes in the buffer.

Removing post-completion behavior in company-files, as it counters this.

Also changing the exit-function argument in company-capf--post-completion
because we /really/ do the same as `completion--done` now.